Move tests, delete obsolete files
